pyproject.toml
[project]
name = "rtfm"
version = "0.1.0"
description = "RTFM Agent - RAG documentation assistant"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    # AI/LLM
    "langchain>=0.3.0",
    "langchain-redis>=0.1.0",
    "langchain-community>=0.3.0",
    "ollama>=0.4.0",
    # Document processing
    "pypdf>=4.3.1",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.3",
    "tiktoken>=0.8.0",
    # Redis
    "redis>=5.2.0",
    "redisvl>=0.15.0",
    # Utilities
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "psutil>=6.1.0",  # For memory monitoring on M4 Air
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.2",
    "ruff>=0.7.0",      # Fast linting (same team as uv!)
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# Optimize for Apple Silicon
index-strategy = "unsafe-best-match"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
# Optional: linting config
line-length = 88
target-version = "py312"
